requirements.txt
absl-py==2.2.2
numpy==1.26.4
openvino-dev==2024.6.0
lap==0.5.12